Rename mvi module to trending-repositories
