+++ app/src/main/kotlin/com/thomaskioko/stargazer/ui/SplashFragment.kt
@@ -8,7 +8,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.thomaskioko.stargazer.R
-import com.thomaskioko.stargazer.navigation.NavigationScreen.RepoListScreen
+import com.thomaskioko.stargazer.navigation.NavigationScreen.TrendingRepositoriesScreen
 import com.thomaskioko.stargazer.navigation.ScreenNavigator
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
@@ -35,7 +35,7 @@ class SplashFragment : Fragment() {
         Handler(Looper.getMainLooper())
             .postDelayed(
                 {
-                    screenNavigator.goToScreen(RepoListScreen)
+                    screenNavigator.goToScreen(TrendingRepositoriesScreen)
                 },
                 SPLASH_DELAY_DURATION
             )
