+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -2,15 +2,16 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
 
     <item
+        android:id="@id/trending_nav_graph"
+        android:enabled="true"
+        android:icon="@drawable/ic_outline_list"
+        android:title="@string/menu_item_trending" />
+
+    <item
         android:id="@id/browse_nav_graph"
         android:enabled="true"
         android:icon="@drawable/ic_list"
         android:title="@string/menu_item_mvvm" />
-    <item
-        android:id="@id/browse_mvi_nav_graph"
-        android:enabled="true"
-        android:icon="@drawable/ic_outline_list"
-        android:title="@string/menu_item_mvi" />
 
     <item
         android:id="@id/bookmark_nav_graph"
