--- features/browse-mvi/src/main/kotlin/com/thomaskioko/stargazer/browse_mvi/model/RepoViewDataModel.kt
@@ -1,14 +1,0 @@
-package com.thomaskioko.stargazer.browse_mvi.model
-
-internal data class RepoViewDataModel(
-    val repoId: Long,
-    val name: String,
-    val description: String?,
-    val userName: String,
-    val stargazersCount: Int,
-    val forksCount: Int,
-    val contributorsUrl: String,
-    val createdDate: String,
-    val updatedDate: String,
-    val isBookmarked: Boolean
-)
