--- features/browse-mvi/src/main/res/navigation/browse_mvi_nav_graph.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@id/browse_mvi_nav_graph"
-    app:startDestination="@id/mviRepoListFragment">
-
-
-    <fragment
-        android:id="@+id/mviRepoListFragment"
-        android:name="com.thomaskioko.stargazer.browse_mvi.ui.MviRepoListFragment"
-        android:label="MviRepoListFragment"
-        tools:layout="@layout/fragment_mvi_repo_list" />
-
-</navigation>
