--- features/browse-mvi/src/test/kotlin/com/thomaskioko/stargazer/browse_mvi/interactor/ViewMockData.kt
@@ -1,37 +1,0 @@
-package com.thomaskioko.stargazer.browse_mvi.interactor
-
-import com.thomaskioko.stargazer.browse_mvi.model.RepoViewDataModel
-import com.thomaskioko.stargazer.db.model.RepoEntity
-
-internal object ViewMockData {
-
-    fun makeRepoViewDataModelList() = listOf(
-        RepoViewDataModel(
-            repoId = 1L,
-            name = "Square",
-            description = "Some cool description about the app",
-            userName = "ninja",
-            stargazersCount = 1,
-            forksCount = 1,
-            contributorsUrl = "",
-            createdDate = "1/11/1900",
-            updatedDate = "1/11/1900",
-            isBookmarked = false
-        )
-    )
-
-    fun makeRepoEntityList() = listOf(
-        RepoEntity(
-            repoId = 1L,
-            name = "Square",
-            description = "Some cool description about the app",
-            userName = "ninja",
-            stargazersCount = 1,
-            forksCount = 1,
-            contributorsUrl = "",
-            createdDate = "1/11/1900",
-            updatedDate = "1/11/1900",
-            isBookmarked = false
-        )
-    )
-}
