+++ features/trending-repositories/src/main/kotlin/com/thomaskioko/stargazer/trending/model/RepoViewDataModel.kt
@@ -1,0 +1,14 @@
+package com.thomaskioko.stargazer.trending.model
+
+internal data class RepoViewDataModel(
+    val repoId: Long,
+    val name: String,
+    val description: String?,
+    val userName: String,
+    val stargazersCount: Int,
+    val forksCount: Int,
+    val contributorsUrl: String,
+    val createdDate: String,
+    val updatedDate: String,
+    val isBookmarked: Boolean
+)
