+++ features/trending-repositories/src/main/res/layout/item_repository.xml
@@ -1,0 +1,103 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <data>
+
+        <import type="android.view.View" />
+
+        <variable
+            name="clickHandler"
+            type="com.thomaskioko.stargazer.trending.ui.RepoItemClick" />
+
+        <variable
+            name="repo"
+            type="com.thomaskioko.stargazer.trending.model.RepoViewDataModel" />
+    </data>
+
+    <FrameLayout
+        android:id="@+id/repo_frame"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginHorizontal="@dimen/grid_0_5"
+        android:layout_marginVertical="@dimen/grid_0_25"
+        android:background="?attr/colorSecondary">
+
+        <com.google.android.material.card.MaterialCardView
+            android:id="@+id/repo_item_cardview"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:clickable="true"
+            android:elevation="@dimen/plane_00"
+            android:focusable="true"
+            android:transitionName="@{@string/repo_card_transition_name(repo.repoId)}"
+            app:cardPreventCornerOverlap="false">
+
+            <androidx.constraintlayout.widget.ConstraintLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:background="?android:attr/selectableItemBackground"
+                android:onClick="@{(v) -> clickHandler.onClick(v, repo.repoId)}">
+
+                <TextView
+                    android:id="@+id/repo_name"
+                    android:layout_width="0dp"
+                    android:layout_height="wrap_content"
+                    android:layout_marginStart="16dp"
+                    android:layout_marginTop="16dp"
+                    android:text="@{repo.name}"
+                    android:textAppearance="?attr/textAppearanceHeadline6"
+                    app:layout_constraintEnd_toStartOf="@+id/stars_count"
+                    app:layout_constraintStart_toStartOf="parent"
+                    app:layout_constraintTop_toTopOf="parent"
+                    tools:text="Square" />
+
+                <TextView
+                    android:id="@+id/repo_description"
+                    android:layout_width="0dp"
+                    android:layout_height="wrap_content"
+                    android:layout_marginStart="16dp"
+                    android:layout_marginTop="8dp"
+                    android:layout_marginBottom="16dp"
+                    android:maxLines="2"
+                    android:text="@{repo.description}"
+                    android:textAppearance="?attr/textAppearanceSubtitle1"
+                    app:layout_constraintBottom_toBottomOf="parent"
+                    app:layout_constraintEnd_toStartOf="@+id/fork_count"
+                    app:layout_constraintStart_toStartOf="parent"
+                    app:layout_constraintTop_toBottomOf="@+id/repo_name"
+                    tools:text="Why are pizza boxes square?" />
+
+                <TextView
+                    android:id="@+id/fork_count"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_marginEnd="16dp"
+                    android:drawableStart="@drawable/ic_fork"
+                    android:gravity="center"
+                    android:textAppearance="?attr/textAppearanceCaption"
+                    app:count="@{repo.forksCount}"
+                    app:layout_constraintEnd_toEndOf="parent"
+                    app:layout_constraintTop_toTopOf="@+id/repo_name"
+                    tools:ignore="UseCompatTextViewDrawableXml"
+                    tools:text="50" />
+
+                <TextView
+                    android:id="@+id/stars_count"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_marginEnd="16dp"
+                    android:drawableStart="@{repo.bookmarked ? @drawable/ic_favorite : @drawable/ic_favorite_black}"
+                    android:gravity="center"
+                    android:textAppearance="?attr/textAppearanceCaption"
+                    app:count="@{repo.stargazersCount}"
+                    app:layout_constraintEnd_toEndOf="parent"
+                    app:layout_constraintTop_toTopOf="@+id/repo_description"
+                    tools:ignore="UseCompatTextViewDrawableXml"
+                    tools:text="150" />
+            </androidx.constraintlayout.widget.ConstraintLayout>
+
+        </com.google.android.material.card.MaterialCardView>
+    </FrameLayout>
+</layout>
