+++ features/trending-repositories/src/main/res/navigation/trending_nav_graph.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@id/trending_nav_graph"
+    app:startDestination="@id/trendingListFragment">
+
+
+    <fragment
+        android:id="@+id/trendingListFragment"
+        android:name="com.thomaskioko.stargazer.trending.ui.TrendingListFragment"
+        android:label="TrendingListFragment"
+        tools:layout="@layout/fragment_trending_list" />
+
+</navigation>
