+++ features/trending-repositories/src/test/kotlin/com/thomaskioko/stargazer/trending/interactor/ViewMockData.kt
@@ -1,0 +1,37 @@
+package com.thomaskioko.stargazer.trending.interactor
+
+import com.thomaskioko.stargazer.trending.model.RepoViewDataModel
+import com.thomaskioko.stargazer.db.model.RepoEntity
+
+internal object ViewMockData {
+
+    fun makeRepoViewDataModelList() = listOf(
+        RepoViewDataModel(
+            repoId = 1L,
+            name = "Square",
+            description = "Some cool description about the app",
+            userName = "ninja",
+            stargazersCount = 1,
+            forksCount = 1,
+            contributorsUrl = "",
+            createdDate = "1/11/1900",
+            updatedDate = "1/11/1900",
+            isBookmarked = false
+        )
+    )
+
+    fun makeRepoEntityList() = listOf(
+        RepoEntity(
+            repoId = 1L,
+            name = "Square",
+            description = "Some cool description about the app",
+            userName = "ninja",
+            stargazersCount = 1,
+            forksCount = 1,
+            contributorsUrl = "",
+            createdDate = "1/11/1900",
+            updatedDate = "1/11/1900",
+            isBookmarked = false
+        )
+    )
+}
