+++ features/trending-repositories/src/test/resources/org.powermock.extensions/configuration.properties
