+++ navigation/src/main/kotlin/com/thomaskioko/stargazer/navigation/ScreenNavigator.kt
@@ -4,15 +4,15 @@ import androidx.navigation.NavController
 import androidx.navigation.Navigator
 import com.thomaskioko.stargazer.actions.MainNavGraphDirections
 import com.thomaskioko.stargazer.navigation.NavigationScreen.BookmarkListScreen
-import com.thomaskioko.stargazer.navigation.NavigationScreen.MviRepoListScreen
+import com.thomaskioko.stargazer.navigation.NavigationScreen.TrendingRepositoriesScreen
 import com.thomaskioko.stargazer.navigation.NavigationScreen.RepoDetailScreen
 import com.thomaskioko.stargazer.navigation.NavigationScreen.RepoListScreen
 import javax.inject.Inject
 import javax.inject.Provider
 
 sealed class NavigationScreen {
+    object TrendingRepositoriesScreen : NavigationScreen()
     object RepoListScreen : NavigationScreen()
-    object MviRepoListScreen : NavigationScreen()
     object BookmarkListScreen : NavigationScreen()
     data class RepoDetailScreen(
         val repoId: Long,
@@ -34,8 +34,8 @@ class ScreenNavigationImpl @Inject constructor(
                 .navigate(MainNavGraphDirections.actionRepoList())
             BookmarkListScreen -> navControllerProvider.get()
                 .navigate(MainNavGraphDirections.actionBookmarkList())
-            MviRepoListScreen -> navControllerProvider.get()
-                .navigate(MainNavGraphDirections.actionMviRepoList())
+            TrendingRepositoriesScreen -> navControllerProvider.get()
+                .navigate(MainNavGraphDirections.actionTrendingRepos())
             is RepoDetailScreen -> navControllerProvider.get()
                 .navigate(
                     MainNavGraphDirections.actionRepoDetails(navigationScreen.repoId),
