--- navigation/src/main/res/navigation/browse_mvi_nav_graph.xml
@@ -1,3 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@id/browse_mvi_nav_graph" />
