+++ navigation/src/main/res/navigation/trending_nav_graph.xml
@@ -1,0 +1,3 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@id/trending_nav_graph" />
